Refactors CAS auth logic
Brandon, Jon, and I identified a previously unsupported use case:
allowing a manuall login user to revert back to using CAS for
Single Sign On. Refactoring the existing logic allows for this
use case to be supported as well as maintaining the change to allow
manual accounts to see the manual login screen by default.