Adds info about SSO/manual accounts to login form
As part of the changes to how manual accounts are being supported,
some additional language was desired for the login form.
Part of this change was including a (direct) link to the CAS
login link for Moodle, so that it is easy to use CAS to login
after logging in with a manual Moodle account (and telling
Moodle to remember your username in a cookie).